requirements.txt
langchain==0.3.9
langchain_anthropic==0.3.0
langchain_community==0.3.8
langchain_core==0.3.21
langchain_openai==0.2.10
langchain_pinecone==0.2.0
langchain_voyageai==0.1.3
pinecone
pydantic==2.10.2
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.39.0
upstash_redis
